Huge refactoring and adding Mac framework and sample app (not working); added a lot of preprocessor #if to get it compiling on Mac without errors

+++ EvernoteSDK/internal/ENCredentials.m
@@ -68,7 +68,7 @@ authenticationToken:(NSString *)authenticationToken
                                    account:self.edamUserId 
                                      error:&error];
     if (!success) {
-        NSLog(@"Error saving to keychain: %@ %d", error, error.code);
+        NSLog(@"Error saving to keychain: %@ %ld", error, (long)error.code);
         return NO;
     } 
     return YES;
